Create tiles component
