+++ web/src/app/simple/styles.module.scss
@@ -10,15 +10,3 @@
   flex-direction: column;
   align-items: center;
 }
-
-.tile {
-  position: relative;
-  background-color: #ddd;
-  border-radius: 20px;
-  height: 1000px;
-  width: 100%;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
