+++ web/src/components/Simple/Tiles/styles.module.scss
@@ -1,0 +1,11 @@
+.tile {
+  position: relative;
+  background-color: #ddd;
+  border-radius: 20px;
+  height: 1000px;
+  width: 100%;
+
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
